requirements.txt
streamlit
requests
pandas
matplotlib
pytesseract
Pillow


